classification/keras/code/requirements.txt
pydot==1.4.2
neptune-tensorflow-keras
neptune_client
numpy==1.23.5
tensorflow==2.11.0
